Compile fix
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20090 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/collide/rapid/csrapid.cpp
@@ -80,7 +80,7 @@ csCollisionPair* csRapidCollideSystem::GetCollisionPairs ()
   return csRapidCollider::GetCollisions ();
 }
 
-int csRapidCollideSystem::GetCollisionPairCount ()
+size_t csRapidCollideSystem::GetCollisionPairCount ()
 {
   return csRapidCollider::numHits;
 }
